requirements.txt
homeassistant>=2024.0.0
jinja2
pyyaml
pydantic
pytest
